create cost sum calculation

+++ checkout-service/src/main/java/com/gca/checkoutservice/logic/OrderItemService.java
@@ -9,5 +9,5 @@ public interface OrderItemService {
     List<OrderItemDto> getAllOrderItemsByOrderId(Integer orderId);
     OrderItemDto addOrderItem(OrderItemDto dto);
     List<OrderItemDto> addOrderIntoOrderItems(Integer orderId, List<Integer> orderItemIds);
-    Integer getCostSumOfOrderItems(List<Integer> orderItemIds) throws JsonProcessingException;
+    Double getCostSumOfOrderItems(List<Integer> orderItemIds) throws JsonProcessingException;
 }
